ユーザー設定コンテキストを分割し設定画面の画像参照を修正。 ユーザー設定画面にWIPアニメカードを追加してコンテキスト構成を整理。
