+++ main/src/pages/userConfig/userConfig.tsx
@@ -1,4 +1,4 @@
-import {useUserConfig} from '../tests/test_page/userConfigContext';
+import {useUserConfig} from '../tests/test_page/hooks/useUserConfig';
 
 import {useMemo} from 'react';
 
@@ -50,6 +50,17 @@ export default function UserConfig() {
           ))}
         </div>
       </section>
+      <div className='relative z-20 m-4 mb-4 mt-10 h-auto w-full max-w-3xl overflow-hidden rounded-2xl border border-white/10 bg-white/5 p-4 shadow-xl backdrop-blur'>
+        <div className='font-mono absolute left-6 top-4 h-24 w-48 rounded-xl border border-white/30 bg-white/20 p-4 text-sm text-white shadow-[0_10px_40px_rgba(0,0,0,0.45)] backdrop-blur'>
+          {'</>'}
+        </div>
+        <div className='font-mono absolute right-8 top-10 h-28 w-52 rounded-xl border border-white/40 bg-white/25 p-4 text-xs text-white shadow-[0_10px_40px_rgba(0,0,0,0.45)] backdrop-blur'>
+          WIP
+        </div>
+        <div className='absolute left-1/2 top-16 h-20 w-40 -translate-x-1/2 transform rounded-xl border border-white/30 bg-white/10 p-4 text-center text-xs text-white shadow-[0_10px_40px_rgba(0,0,0,0.35)] backdrop-blur'>
+          Building...
+        </div>
+      </div>
     </AppLayout>
   );
 }
